+++ animals/fosterprogram/permissions.py
@@ -1,0 +1,6 @@
+from rest_framework.permissions import BasePermission
+
+
+class IsUserOrAdmin(BasePermission):
+    def has_permission(self, request, view):
+        return request.user.is_authenticated
