+++ animals/fosterprogram/serializers.py
@@ -1,3 +1,8 @@
 from rest_framework import serializers
+from .models import FosterApplication
 
-# Define FosterApplicationSerializer here
+class FosterApplicationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = FosterApplication
+        fields = '__all__'
+        read_only_fields = ['user', 'submitted_at']
